refactor: factor out color chooser
