+++ packages/apps/plugins/plugin-space/src/types.ts
@@ -39,7 +39,6 @@ export type ObjectViewer = {
   spaceKey: PublicKey;
   objectId: string;
   lastSeen: number;
-  color?: string;
 };
 
 export type PluginState = {
