+++ packages/ui/react-ui-theme/src/util/index.ts
@@ -2,4 +2,5 @@
 // Copyright 2022 DXOS.org
 //
 
+export * from './color';
 export * from './mx';
